fix(adapter): Resolves a stop token issue for Local LLM Adapter (#1012)
